Summarise continuous variables by group

Summarise continuous variables by group. Non-numeric variables will be dropped.

Usage

summ_by(d, cols = NULL, ..., pct = c(0.25, 0.75), xname = "")

Arguments

d

<dfr> A data frame

cols

<chr> A vector of column names to select

...

<var> Columns to group by as unquoted variable names

pct

<num> A vector of two indicating the percentiles to compute

xname

<chr> Characters to omit in output column names

Value

A data frame of summarised variables

Examples

mtcars |> summ_by()
#> Dropped:  
#> Adding missing grouping variables: `name`
#> # A tibble: 11 × 10
#>    name      n   nNA    Mean    Med      SD   Min    P25    P75    Max
#>    <chr> <int> <int>   <dbl>  <dbl>   <dbl> <dbl>  <dbl>  <dbl>  <dbl>
#>  1 am       32     0   0.406   0      0.499  0      0      1      1   
#>  2 carb     32     0   2.81    2      1.62   1      2      4      8   
#>  3 cyl      32     0   6.19    6      1.79   4      4      8      8   
#>  4 disp     32     0 231.    196.   124.    71.1  121.   326    472   
#>  5 drat     32     0   3.60    3.70   0.535  2.76   3.08   3.92   4.93
#>  6 gear     32     0   3.69    4      0.738  3      3      4      5   
#>  7 hp       32     0 147.    123     68.6   52     96.5  180    335   
#>  8 mpg      32     0  20.1    19.2    6.03  10.4   15.4   22.8   33.9 
#>  9 qsec     32     0  17.8    17.7    1.79  14.5   16.9   18.9   22.9 
#> 10 vs       32     0   0.438   0      0.504  0      0      1      1   
#> 11 wt       32     0   3.22    3.32   0.978  1.51   2.58   3.61   5.42
mtcars |> summ_by(pct=c(0.1,0.9))
#> Dropped:  
#> Adding missing grouping variables: `name`
#> # A tibble: 11 × 10
#>    name      n   nNA    Mean    Med      SD   Min   P10    P90    Max
#>    <chr> <int> <int>   <dbl>  <dbl>   <dbl> <dbl> <dbl>  <dbl>  <dbl>
#>  1 am       32     0   0.406   0      0.499  0     0      1      1   
#>  2 carb     32     0   2.81    2      1.62   1     1      4      8   
#>  3 cyl      32     0   6.19    6      1.79   4     4      8      8   
#>  4 disp     32     0 231.    196.   124.    71.1  80.6  396    472   
#>  5 drat     32     0   3.60    3.70   0.535  2.76  3.01   4.21   4.93
#>  6 gear     32     0   3.69    4      0.738  3     3      5      5   
#>  7 hp       32     0 147.    123     68.6   52    66    244.   335   
#>  8 mpg      32     0  20.1    19.2    6.03  10.4  14.3   30.1   33.9 
#>  9 qsec     32     0  17.8    17.7    1.79  14.5  15.5   20.0   22.9 
#> 10 vs       32     0   0.438   0      0.504  0     0      1      1   
#> 11 wt       32     0   3.22    3.32   0.978  1.51  1.96   4.05   5.42
mtcars |> summ_by("mpg")
#> Dropped:  
#> Adding missing grouping variables: `name`
#> # A tibble: 1 × 10
#>   name      n   nNA  Mean   Med    SD   Min   P25   P75   Max
#>   <chr> <int> <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 mpg      32     0  20.1  19.2  6.03  10.4  15.4  22.8  33.9
mtcars |> summ_by("mpg",vs)
#> # A tibble: 2 × 10
#>      vs     n   nNA  Mean   Med    SD   Min   P25   P75   Max
#>   <dbl> <int> <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1     0    18     0  16.6  15.6  3.86  10.4  14.8  19.1  26  
#> 2     1    14     0  24.6  22.8  5.38  17.8  21.4  29.6  33.9
mtcars |> summ_by("mpg",vs,am)
#> # A tibble: 4 × 11
#> # Groups:   vs [2]
#>      vs    am     n   nNA  Mean   Med    SD   Min   P25   P75   Max
#>   <dbl> <dbl> <int> <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1     0     0    12     0  15.0  15.2  2.77  10.4  14.0  16.6  19.2
#> 2     0     1     6     0  19.8  20.4  4.01  15    16.8  21    26  
#> 3     1     0     7     0  20.7  21.4  2.47  17.8  18.6  22.2  24.4
#> 4     1     1     7     0  28.4  30.4  4.76  21.4  25.0  31.4  33.9
mtcars |> summ_by(c("mpg","disp"))
#> Dropped:  
#> Adding missing grouping variables: `name`
#> # A tibble: 2 × 10
#>   name      n   nNA  Mean   Med     SD   Min   P25   P75   Max
#>   <chr> <int> <int> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1 disp     32     0 231.  196.  124.    71.1 121.  326   472  
#> 2 mpg      32     0  20.1  19.2   6.03  10.4  15.4  22.8  33.9
mtcars |> summ_by(c("mpg","disp"),vs)
#> # A tibble: 2 × 19
#>      vs mpg_n mpg_nNA mpg_Mean mpg_Med mpg_SD mpg_Min mpg_P25 mpg_P75 mpg_Max
#>   <dbl> <int>   <int>    <dbl>   <dbl>  <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1     0    18       0     16.6    15.6   3.86    10.4    14.8    19.1    26  
#> 2     1    14       0     24.6    22.8   5.38    17.8    21.4    29.6    33.9
#> # ℹ 9 more variables: disp_n <int>, disp_nNA <int>, disp_Mean <dbl>,
#> #   disp_Med <dbl>, disp_SD <dbl>, disp_Min <dbl>, disp_P25 <dbl>,
#> #   disp_P75 <dbl>, disp_Max <dbl>
mtcars |> summ_by(c("mpg","disp"),vs,xname="mpg_")
#> # A tibble: 2 × 19
#>      vs     n   nNA  Mean   Med    SD   Min   P25   P75   Max disp_n disp_nNA
#>   <dbl> <int> <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <int>    <int>
#> 1     0    18     0  16.6  15.6  3.86  10.4  14.8  19.1  26       18        0
#> 2     1    14     0  24.6  22.8  5.38  17.8  21.4  29.6  33.9     14        0
#> # ℹ 7 more variables: disp_Mean <dbl>, disp_Med <dbl>, disp_SD <dbl>,
#> #   disp_Min <dbl>, disp_P25 <dbl>, disp_P75 <dbl>, disp_Max <dbl>
# grouping without column selection is possible but rarely useful in large dataset
mtcars |> summ_by(NULL,vs)
#> # A tibble: 2 × 91
#>      vs mpg_n mpg_nNA mpg_Mean mpg_Med mpg_SD mpg_Min mpg_P25 mpg_P75 mpg_Max
#>   <dbl> <int>   <int>    <dbl>   <dbl>  <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#> 1     0    18       0     16.6    15.6   3.86    10.4    14.8    19.1    26  
#> 2     1    14       0     24.6    22.8   5.38    17.8    21.4    29.6    33.9
#> # ℹ 81 more variables: cyl_n <int>, cyl_nNA <int>, cyl_Mean <dbl>,
#> #   cyl_Med <dbl>, cyl_SD <dbl>, cyl_Min <dbl>, cyl_P25 <dbl>, cyl_P75 <dbl>,
#> #   cyl_Max <dbl>, disp_n <int>, disp_nNA <int>, disp_Mean <dbl>,
#> #   disp_Med <dbl>, disp_SD <dbl>, disp_Min <dbl>, disp_P25 <dbl>,
#> #   disp_P75 <dbl>, disp_Max <dbl>, hp_n <int>, hp_nNA <int>, hp_Mean <dbl>,
#> #   hp_Med <dbl>, hp_SD <dbl>, hp_Min <dbl>, hp_P25 <dbl>, hp_P75 <dbl>,
#> #   hp_Max <dbl>, drat_n <int>, drat_nNA <int>, drat_Mean <dbl>, …